+++ src/plugins/error-messages.js
@@ -148,7 +148,10 @@ const CODES = {
 
     // plugins/validate.js
     VD1: 'Sub-schema not found, does the schemaPath exists in your schema?',
-    VD2: 'object does not match schema'
+    VD2: 'object does not match schema',
+
+    // plugins/in-memory.js
+    IM1: 'InMemory: Memory-Adapter must be added. Use RxDB.plugin(require(\'pouchdb-adapter-memory\'));'
 };
 
 export const rxdb = true;